pyproject.toml
[tool.poetry]
name = "repochat"
version = "0.1.0"
description = ""
authors = ["Mike Fischer <mike@theelectricrambler.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
streamlit = "^1.31.1"
chromadb = "^0.4.22"
tiktoken = "^0.6.0"
sentence-transformers = "^2.3.1"
termcolor = "^2.4.0"
langchain-community = "^0.0.19"
pyyaml = "^6.0.1"
gitpython = "^3.1.41"
openai = "^1.12.0"
huggingface-hub = "^0.20.3"
htmltemplate = "^2.2.1"
pypdf = "^4.0.2"
textsplitter = "^1.0.3"
langchain = "0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
